+++ templates/node/CHANGELOG.md
@@ -1,5 +1,14 @@
 # node-name
 
+## 0.0.16
+
+### Patch Changes
+
+- Added examples and create project' command to include examples and 'create node' command with options for type ('module' or 'class') and template ('class' or 'ui')
+- Updated dependencies
+  - @nanoservice-ts/runner@0.1.18
+  - @nanoservice-ts/shared@0.0.8
+
 ## 0.0.15
 
 ### Patch Changes